requirements.txt
discord_webhook==1.1.0
Flask==2.2.3
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
speedtest==0.0.1
speedtest_cli==2.1.3
